refactor: asyncActions.ts 에서 따로 관리 되던 비동기 액션 slice.ts 에서 관리하는 방식으로 수정
  - refactor: createAsyncAction() 하기 전에 따로 객체로 액션 한번 더 만들어주는 작업 생략하고, 바로 createAsyncAction 안에 파라미터로 바로 액션 선원과 동시에 넣어주는 방식으로 수정

  - refactor: slice 에서 관리되면 reducer initial state, types.ts 에서 관리하는 방식으로 수정
